Rename to OpenAITextGenerationService and samples
