+++ java/aiservices/openai/src/main/java/com/microsoft/semantickernel/aiservices/openai/OpenAITextGenerationService.java
@@ -1,4 +1,4 @@
-package com.microsoft.semantickernel.aiservices.azureopenai;
+package com.microsoft.semantickernel.aiservices.openai;
 
 import com.azure.ai.openai.OpenAIAsyncClient;
 import com.azure.ai.openai.models.Choice;
@@ -19,7 +19,7 @@ import javax.annotation.Nullable;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public class AzureOpenAITextGenerationService implements TextGenerationService {
+public class OpenAITextGenerationService implements TextGenerationService {
 
     private final OpenAIAsyncClient client;
     private final Map<String, ContextVariable<?>> attributes;
@@ -33,7 +33,7 @@ public class AzureOpenAITextGenerationService implements TextGenerationService {
     /// <param name="modelId">Azure OpenAI model id, see https://learn.microsoft.com/azure/cognitive-services/openai/how-to/create-resource</param>
     /// <param name="httpClient">Custom <see cref="HttpClient"/> for HTTP requests.</param>
     /// <param name="loggerFactory">The <see cref="ILoggerFactory"/> to use for logging. If null, no logging will be performed.</param>
-    public AzureOpenAITextGenerationService(
+    public OpenAITextGenerationService(
         OpenAIAsyncClient client,
         String modelId) {
         this.client = client;
@@ -131,8 +131,8 @@ public class AzureOpenAITextGenerationService implements TextGenerationService {
             return this;
         }
 
-        public AzureOpenAITextGenerationService build() {
-            return new AzureOpenAITextGenerationService(
+        public OpenAITextGenerationService build() {
+            return new OpenAITextGenerationService(
                 this.client,
                 this.modelId
             );
