+++ java/samples/sample-code/src/main/java/com/microsoft/semantickernel/samples/syntaxexamples/Example06_TemplateLanguage.java
@@ -6,7 +6,7 @@ import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.credential.KeyCredential;
 import com.microsoft.semantickernel.DefaultKernel;
 import com.microsoft.semantickernel.Kernel;
-import com.microsoft.semantickernel.aiservices.azureopenai.AzureOpenAITextGenerationService;
+import com.microsoft.semantickernel.aiservices.openai.OpenAITextGenerationService;
 import com.microsoft.semantickernel.exceptions.ConfigurationException;
 import com.microsoft.semantickernel.orchestration.PromptExecutionSettings.Builder;
 import com.microsoft.semantickernel.plugin.KernelFunctionFactory;
@@ -43,26 +43,23 @@ public class Example06_TemplateLanguage {
 
         System.out.println("======== TemplateLanguage ========");
 
-        TextGenerationService textGenerationService;
+        OpenAIAsyncClient client;
 
         if (USE_AZURE_CLIENT) {
-            OpenAIAsyncClient client = new OpenAIClientBuilder()
+            client = new OpenAIClientBuilder()
                 .credential(new AzureKeyCredential(CLIENT_KEY))
                 .endpoint(CLIENT_ENDPOINT)
                 .buildAsyncClient();
+        } else {
+            client = new OpenAIClientBuilder()
+                .credential(new KeyCredential(CLIENT_KEY))
+                .buildAsyncClient();
+        }
 
-            textGenerationService = AzureOpenAITextGenerationService.builder()
+        TextGenerationService textGenerationService = OpenAITextGenerationService.builder()
                 .withOpenAIAsyncClient(client)
                 .withModelId("text-davinci-003")
                 .build();
-        } else {
-            OpenAIAsyncClient client = new OpenAIClientBuilder()
-                .credential(new KeyCredential(CLIENT_KEY))
-                .buildAsyncClient();
-
-            // TODO: Add support for OpenAI API
-            textGenerationService = null;
-        }
 
         Kernel kernel = new DefaultKernel.Builder()
             .withDefaultAIService(TextGenerationService.class, textGenerationService)
